Refactoring: kind/union -> class Type/std.variant + fixes
class Cell: `int kind` and `enum` of type values have been replaced by a
`class Type` and the use of std.variant.Variant.

This should make it easier to extend Stacky with custom types.

Initially, Stacky was meant as a D prototype for a future C/C++
implementation. I am giving up on that idea.

With the `Type` class we may get multi methods and nice stuffs (like type
checking the stack effects and procs arity) for using stacky as a target
language for a higher level, functional other one.

Fixes:
- an `ExeCtrl` type with one value for stacky to know that we exited a nested
  proc, so that it returns from its eval function.
  It is added at the end of the code array of Word procs.

- `not` word was not pushing a boolean on the stack: spotted as a type error
  by `std.variant`.