More typing, symbol contexts, fixes, better repl.
"->" symbol for function application, fixed "*" (the product).

Symbols can be evaluated differently according to a context: we use some
regex and an array of delegate to identify symbols that match and the
modified behaviour to apply.
This allows us to drop the '/' before constructor names and type variables to
start with colons ':'.

Fixed the `for` loop word. Works as expected.

In the `repl` we now check if the procedure curl braces are balanced to
buffer the input and properly evaluate multi-line inputs.